fix: resolve React Fast Refresh linting error by splitting RoomLayout context and provider
