+++ frontend/src/features/collab/context/RoomLayoutContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export interface RoomLayoutState {
+  questionCollapsed: boolean;
+  chatCollapsed: boolean;
+  toggleQuestion: () => void;
+  toggleChat: () => void;
+}
+
+export const RoomLayoutContext = createContext<RoomLayoutState | undefined>(undefined);
